Update structure of composition::supergraph::config module
