+++ src/composition/supergraph/config/error/mod.rs
@@ -1,0 +1,4 @@
+//! This module errors related to resolving SupergraphConfigs that are also shared across multiple submodules of [`rover::composition::supergraph::config`]
+
+mod subgraph;
+pub use subgraph::*;
