+++ src/composition/supergraph/config/unresolved/mod.rs
@@ -1,0 +1,7 @@
+//! Provides objects that represent an unresolved state of a supergraph config.
+
+mod subgraph;
+mod supergraph;
+
+pub use subgraph::*;
+pub use supergraph::*;
