+++ src/composition/supergraph/config/unresolved/supergraph.rs
@@ -1,32 +1,15 @@
 //! Provides tooling to resolve subgraphs, fully or lazily
 use std::collections::BTreeMap;
 
-use apollo_federation_types::config::{
-    FederationVersion, SchemaSource, SubgraphConfig, SupergraphConfig,
-};
+use apollo_federation_types::config::{FederationVersion, SubgraphConfig};
 use buildstructor::buildstructor;
 use camino::Utf8PathBuf;
 use derive_getters::Getters;
-use futures::{
-    stream::{self, StreamExt},
-    TryFutureExt,
-};
-use itertools::Itertools;
-use thiserror::Error;
 
-use crate::utils::effect::{
-    fetch_remote_subgraph::FetchRemoteSubgraph, introspect::IntrospectSubgraph,
-};
-
-use self::subgraph::{
-    FullyResolvedSubgraph, LazilyResolvedSubgraph, ResolveSubgraphError, UnresolvedSubgraph,
-};
-
-use super::ResolveSupergraphConfigError;
-
-pub mod subgraph;
+use super::UnresolvedSubgraph;
 
 /// Object that represents a [`SupergraphConfig`] that requires resolution
+#[derive(Getters)]
 pub struct UnresolvedSupergraphConfig {
     origin_path: Option<Utf8PathBuf>,
     subgraphs: BTreeMap<String, UnresolvedSubgraph>,
@@ -55,227 +38,6 @@ impl UnresolvedSupergraphConfig {
     }
 }
 
-/// Represents a [`SupergraphConfig`] that has a known [`FederationVersion`] and
-/// its subgraph [`SchemaSource`]s reduced to [`SchemaSource::Sdl`]
-#[derive(Clone, Debug, Eq, PartialEq, Getters)]
-pub struct FullyResolvedSupergraphConfig {
-    origin_path: Option<Utf8PathBuf>,
-    subgraphs: BTreeMap<String, FullyResolvedSubgraph>,
-    federation_version: FederationVersion,
-}
-
-impl From<FullyResolvedSupergraphConfig> for SupergraphConfig {
-    fn from(value: FullyResolvedSupergraphConfig) -> Self {
-        let subgraphs = value
-            .subgraphs
-            .into_iter()
-            .map(|(name, subgraph)| (name, subgraph.into()))
-            .collect();
-        SupergraphConfig::new(subgraphs, Some(value.federation_version))
-    }
-}
-
-impl FullyResolvedSupergraphConfig {
-    /// Resolves an [`UnresolvedSupergraphConfig`] into a [`FullyResolvedSupergraphConfig`]
-    /// by resolving the individual subgraphs concurrently and calculating the [`FederationVersion`]
-    pub async fn resolve(
-        introspect_subgraph_impl: &impl IntrospectSubgraph,
-        fetch_remote_subgraph_impl: &impl FetchRemoteSubgraph,
-        supergraph_config_root: Option<&Utf8PathBuf>,
-        unresolved_supergraph_config: UnresolvedSupergraphConfig,
-    ) -> Result<FullyResolvedSupergraphConfig, ResolveSupergraphConfigError> {
-        let subgraphs = stream::iter(unresolved_supergraph_config.subgraphs.into_iter().map(
-            |(name, unresolved_subgraph)| {
-                FullyResolvedSubgraph::resolve(
-                    introspect_subgraph_impl,
-                    fetch_remote_subgraph_impl,
-                    supergraph_config_root,
-                    unresolved_subgraph,
-                )
-                .map_ok(|result| (name, result))
-            },
-        ))
-        .buffer_unordered(50)
-        .collect::<Vec<Result<(String, FullyResolvedSubgraph), ResolveSubgraphError>>>()
-        .await;
-        let (subgraphs, errors): (
-            Vec<(String, FullyResolvedSubgraph)>,
-            Vec<ResolveSubgraphError>,
-        ) = subgraphs.into_iter().partition_result();
-        if errors.is_empty() {
-            let subgraphs = BTreeMap::from_iter(subgraphs);
-            let federation_version = Self::resolve_federation_version(
-                unresolved_supergraph_config.federation_version,
-                &mut subgraphs.iter(),
-            )?;
-            Ok(FullyResolvedSupergraphConfig {
-                origin_path: unresolved_supergraph_config.origin_path,
-                subgraphs,
-                federation_version,
-            })
-        } else {
-            Err(ResolveSupergraphConfigError::ResolveSubgraphs(errors))
-        }
-    }
-
-    fn resolve_federation_version<'a>(
-        specified_federation_version: Option<FederationVersion>,
-        subgraphs: &'a mut impl Iterator<Item = (&'a String, &'a FullyResolvedSubgraph)>,
-    ) -> Result<FederationVersion, ResolveSupergraphConfigError> {
-        let fed_two_subgraphs = subgraphs
-            .filter_map(|(subgraph_name, subgraph)| {
-                if *subgraph.is_fed_two() {
-                    Some(subgraph_name.to_string())
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>();
-        let contains_fed_two_subgraphs = !fed_two_subgraphs.is_empty();
-        match specified_federation_version {
-            Some(specified_federation_version) => {
-                if specified_federation_version.is_fed_one() {
-                    if contains_fed_two_subgraphs {
-                        Err(ResolveSupergraphConfigError::FederationVersionMismatch {
-                            specified_federation_version,
-                            subgraph_names: fed_two_subgraphs,
-                        })
-                    } else {
-                        Ok(specified_federation_version)
-                    }
-                } else {
-                    Ok(specified_federation_version)
-                }
-            }
-            None => Ok(FederationVersion::LatestFedTwo),
-        }
-    }
-}
-
-/// Represents a [`SupergraphConfig`] where all its [`SchemaSource::File`] subgraphs have
-/// known and valid file paths relative to a supergraph config file (or working directory of the
-/// program, if the supergraph config is piped into stdin)
-#[derive(Clone, Debug, Eq, PartialEq, Getters)]
-pub struct LazilyResolvedSupergraphConfig {
-    origin_path: Option<Utf8PathBuf>,
-    subgraphs: BTreeMap<String, LazilyResolvedSubgraph>,
-    federation_version: Option<FederationVersion>,
-}
-
-impl LazilyResolvedSupergraphConfig {
-    /// Resolves an [`UnresolvedSupergraphConfig`] into a [`LazilyResolvedSupergraphConfig`] by
-    /// making sure any internal file paths are correct
-    pub async fn resolve(
-        supergraph_config_root: &Utf8PathBuf,
-        unresolved_supergraph_config: UnresolvedSupergraphConfig,
-    ) -> Result<LazilyResolvedSupergraphConfig, Vec<ResolveSubgraphError>> {
-        let subgraphs = stream::iter(unresolved_supergraph_config.subgraphs.into_iter().map(
-            |(name, unresolved_subgraph)| async {
-                let result =
-                    LazilyResolvedSubgraph::resolve(supergraph_config_root, unresolved_subgraph)?;
-                Ok((name, result))
-            },
-        ))
-        .buffer_unordered(50)
-        .collect::<Vec<Result<(String, LazilyResolvedSubgraph), ResolveSubgraphError>>>()
-        .await;
-        let (subgraphs, errors): (
-            Vec<(String, LazilyResolvedSubgraph)>,
-            Vec<ResolveSubgraphError>,
-        ) = subgraphs.into_iter().partition_result();
-        if errors.is_empty() {
-            Ok(LazilyResolvedSupergraphConfig {
-                origin_path: unresolved_supergraph_config.origin_path,
-                subgraphs: BTreeMap::from_iter(subgraphs),
-                federation_version: unresolved_supergraph_config.federation_version,
-            })
-        } else {
-            Err(errors)
-        }
-    }
-}
-
-impl From<LazilyResolvedSupergraphConfig> for SupergraphConfig {
-    fn from(value: LazilyResolvedSupergraphConfig) -> Self {
-        let subgraphs = BTreeMap::from_iter(
-            value
-                .subgraphs
-                .into_iter()
-                .map(|(name, subgraph)| (name, subgraph.into())),
-        );
-        SupergraphConfig::new(subgraphs, value.federation_version)
-    }
-}
-
-/// Error that occurs when a subgraph schema source is invalid
-#[derive(Error, Debug)]
-#[error("Invalid schema source: {:?}", .schema_source)]
-pub struct InvalidSchemaSource {
-    schema_source: SchemaSource,
-}
-
-/// Object that contains the completed set of subgraphs resolved to their SDLs
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct FullyResolvedSubgraphs {
-    subgraphs: BTreeMap<String, String>,
-}
-
-impl FullyResolvedSubgraphs {
-    #[cfg(test)]
-    pub fn new(subgraphs: BTreeMap<String, String>) -> FullyResolvedSubgraphs {
-        FullyResolvedSubgraphs { subgraphs }
-    }
-
-    /// Used to upsert a fully resolved subgraph into this object's definitions
-    pub fn upsert_subgraph(&mut self, name: String, schema: String) {
-        self.subgraphs.insert(name, schema);
-    }
-
-    /// Removes a subgraph from this object's definitions
-    pub fn remove_subgraph(&mut self, name: &str) {
-        self.subgraphs.remove(name);
-    }
-}
-
-impl TryFrom<SupergraphConfig> for FullyResolvedSubgraphs {
-    type Error = Vec<InvalidSchemaSource>;
-    fn try_from(value: SupergraphConfig) -> Result<Self, Self::Error> {
-        let mut errors = Vec::new();
-        let mut subgraph_sdls = BTreeMap::new();
-        for (name, subgraph_config) in value.into_iter() {
-            if let SchemaSource::Sdl { sdl } = subgraph_config.schema {
-                subgraph_sdls.insert(name, sdl);
-            } else {
-                errors.push(InvalidSchemaSource {
-                    schema_source: subgraph_config.schema,
-                });
-            }
-        }
-        if errors.is_empty() {
-            Ok(FullyResolvedSubgraphs {
-                subgraphs: subgraph_sdls,
-            })
-        } else {
-            Err(errors)
-        }
-    }
-}
-
-impl From<FullyResolvedSubgraphs> for SupergraphConfig {
-    fn from(value: FullyResolvedSubgraphs) -> Self {
-        let subgraphs = BTreeMap::from_iter(value.subgraphs.into_iter().map(|(name, sdl)| {
-            (
-                name,
-                SubgraphConfig {
-                    routing_url: None,
-                    schema: SchemaSource::Sdl { sdl },
-                },
-            )
-        }));
-        SupergraphConfig::new(subgraphs, None)
-    }
-}
-
 #[cfg(test)]
 mod tests {
 
@@ -293,16 +55,17 @@ mod tests {
     use speculoos::prelude::*;
 
     use crate::{
-        composition::supergraph::config::ResolveSupergraphConfigError,
+        composition::supergraph::config::{
+            full::{FullyResolvedSubgraph, FullyResolvedSupergraphConfig},
+            lazy::{LazilyResolvedSubgraph, LazilyResolvedSupergraphConfig},
+            resolver::ResolveSupergraphConfigError,
+            scenario::*,
+            unresolved::UnresolvedSupergraphConfig,
+        },
         utils::effect::{
             fetch_remote_subgraph::{MockFetchRemoteSubgraph, RemoteSubgraph},
             introspect::MockIntrospectSubgraph,
         },
-    };
-
-    use super::{
-        subgraph::{scenario::*, FullyResolvedSubgraph, LazilyResolvedSubgraph},
-        FullyResolvedSupergraphConfig, LazilyResolvedSupergraphConfig, UnresolvedSupergraphConfig,
     };
 
     #[fixture]
@@ -632,10 +395,10 @@ mod tests {
                     .build(),
             ),
         ]);
-        assert_that!(resolved_supergraph_config.subgraphs).is_equal_to(expected_subgraphs);
+        assert_that!(resolved_supergraph_config.subgraphs()).is_equal_to(&expected_subgraphs);
 
-        assert_that!(resolved_supergraph_config.federation_version)
-            .is_equal_to(expected_federation_version);
+        assert_that!(resolved_supergraph_config.federation_version())
+            .is_equal_to(&expected_federation_version);
 
         Ok(())
     }
@@ -878,7 +641,7 @@ mod tests {
         .await;
         let resolved_supergraph_config = assert_that!(result).is_ok().subject;
         // fed version is the default, since none provided
-        assert_that!(resolved_supergraph_config.federation_version).is_none();
+        assert_that!(resolved_supergraph_config.federation_version().as_ref()).is_none();
 
         let expected_subgraphs = BTreeMap::from_iter([
             (
@@ -926,7 +689,7 @@ mod tests {
             ),
         ]);
 
-        assert_that!(resolved_supergraph_config.subgraphs).is_equal_to(expected_subgraphs);
+        assert_that!(resolved_supergraph_config.subgraphs()).is_equal_to(&expected_subgraphs);
 
         Ok(())
     }
